Hide deleted comments in additional places
Apparently i forgot to hide deleted comments in some places, namely _the homepage_ ...
This hides them on the homepage, the release viewer, and the generic asset editor.

In the future we might want to show them again for moderators, but for now its better to just hide them unconditionally imo.