Fixes filedrop filename encoding issue. (#44)
This modifies the jquery-filedrop code, as it did not adhere to the specification with their implementation, which caused a mismatch on how uploaded files are encoded between their version and a native file upload.

This also fixes some issues with the download system, and the `none` cdn that either were never checked, or introduced by previous fixes.