Add mod contributions to 'by user' mod lists (#46)
This change adds the mods a user is on the team of to their user page, as well as their home page.

This also changes the heading above those sections from "[your mods]" to "[Mods you've contributed to]" to make it clear that the user did not necessarily create these mods on their own.

Splittin up these lists into "created" and "contributed to" would be an option, but for now I went with a simple solution.