mod search pagination / endless scrolling
This also changes a few other things, adds indices to tables, changes nullability of some columns and in general changes some approaches.

It does not currently replicate the ranking for text searches, which is really difficult  with all other other constraints.
It does not rely on LIMIT+OFFSET queries, but on a cursor based approach for reasonable performance in most cases.

It does automatic endess scrolling and has fallback for devices that don't support it. There might be compatibility issues with async/await depending on what our target platforms are.